Rename file from asset_urls to asset_paths
