+++ lib/munge.rb
@@ -12,7 +12,7 @@ require "munge/util/path"
 require "munge/util/symbol_hash"
 require "munge/util/config"
 require "munge/helper/asset_tags"
-require "munge/helper/asset_urls"
+require "munge/helper/asset_paths"
 require "munge/helper/capture"
 require "munge/helper/find"
 require "munge/helper/link"
