+++ lib/munge/helper/asset_paths.rb
